+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
++ Bug in `asli download --lsm`, added `-f/--filename` option to CLI.
+
 ## 0.2.0 - 2026-04-02
 
 ### Breaking changes